Syscall Enum verschoben
Enum verschoben, weil jetzt noch zusätzliche Methoden dazu kamen für besseres Handling des Enums
