+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "usrlib"
-version = "0.2.3"
+version = "1.1.1"
 edition = "2021"
 authors = [
     "Michael Schöttner <michael.schoettner@hhu.de>",
